+++ bin/replicante/src/components/webui/cluster/mod.rs
@@ -2,8 +2,10 @@ mod agents;
 mod discovery;
 mod events;
 mod meta;
+mod nodes;
 
 pub use self::agents::Agents;
 pub use self::discovery::Discovery;
 pub use self::events::Events;
 pub use self::meta::Meta;
+pub use self::nodes::Nodes;