+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Cluster specific agents API (`/cluster/:id/agents`).
 - Cluster specific events API (`/cluster/:id/events`).
+- Cluster specific nodes API (`/cluster/:id/nodes`).
 - HTTP Discovery.
 
 ### Changed